Folder Reconstruction, Part 2
Applying all of the folder reconstruction from new-app and applying it to scenario_editor.  server.js and the package.json are maintained. ``npm start`` by default normally calls ``react-scripts start``, now it will call ``server.js`` in order to run it with express. It's aliased to ``npm dev` as it was before.

TL;DR If you want to look at the web pages like before, call ``npm dev`` now.